관리자 펜션관리,회원관리 수정. 선택한 것 없을시 안내창 띄우기. clientList.jsp, pensionList.jsp 수정.

+++ web/views/master/pensionList.jsp
@@ -102,29 +102,35 @@
 		        	
 		        	//삭제버튼 동작
 		        	$("#deleteBtn").click(function(){
-		        		if(confirm("정말로 삭제하시겠습니까?")){
-							var tdArr = new Array();
-							var checkbox = $("input[name=selected]:checked");
-							
-							// 체크된 체크박스 값을 가져온다
-							checkbox.each(function(i) {
-								// checkbox.parent() : checkbox의 부모는 <td>이다.
-								// checkbox.parent().parent() : <td>의 부모이므로 <tr>이다.
-								var tr = checkbox.parent().parent().eq(i);
-								var td = tr.children();
+		        		var checkbox = $("input[name=selected]:checked");
+		        		
+		        		if(checkbox.length==0){
+		        			alert("삭제할 펜션을 선택해주세요.");
+		        		}
+		        		else{
+			        		if(confirm("정말로 삭제하시겠습니까?")){
+								var tdArr = new Array();
 								
-								// td.eq(0)은 체크박스 이므로  td.eq(1)의 값부터 가져온다.
-								var pensionCode = td.eq(1).children().text().trim();
+								// 체크된 체크박스 값을 가져온다
+								checkbox.each(function(i) {
+									// checkbox.parent() : checkbox의 부모는 <td>이다.
+									// checkbox.parent().parent() : <td>의 부모이므로 <tr>이다.
+									var tr = checkbox.parent().parent().eq(i);
+									var td = tr.children();
+									
+									// td.eq(0)은 체크박스 이므로  td.eq(1)의 값부터 가져온다.
+									var pensionCode = td.eq(1).children().text().trim();
 
-								console.log("펜션코드 : "+pensionCode);
+									console.log("펜션코드 : "+pensionCode);
+									
+									// 가져온 값을 배열에 담는다.
+									tdArr.push(pensionCode);
+								});
+								console.log(tdArr);
 								
-								// 가져온 값을 배열에 담는다.
-								tdArr.push(pensionCode);
-							});
-							console.log(tdArr);
-							
-							//삭제할 고객들을 서블릿으로 보내기
-							location.href="<%=request.getContextPath()%>/master/pensionDelete?delPensionList="+tdArr;
+								//삭제할 고객들을 서블릿으로 보내기
+								location.href="<%=request.getContextPath()%>/master/pensionDelete?delPensionList="+tdArr;
+			        		}		        			
 		        		}
 					});
